Move NavigationControl to a component
